backend/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==1.11.26
django-rq==1.3.1
djangorestframework==3.10.3
idna==2.8
psycopg2==2.8.4
pytz==2019.3
redis==3.3.11
requests==2.22.0
rq==0.13.0
urllib3==1.25.7
